science32: Refactor the assembler, maybe fix right shift & a few bugs.
This is leading up to the really scary change: using System.Buffer.BlockCopy for memory accesses to stop torturing the memory allocator.

I am also pretty sure this will set the stage for implementing the M extension also, since we now have a clear idea of how exactly that is to be done.